MAJOR CHANGE -- General knowledge
generalknowledge -- a class that stores and analyzes knowledge expressed as nodes, with directed and nondirected relations between them, and with attributes attached to them. 

Calls to general knowledge class can be done in several ways:
1) By typing the command "generalknowledge," "gk" --- which calls up a prompt.
Command syntax as follows:
NODE1,NODE2,NODE2  =  creates new nodes
RelationName:RelationType;Content = create a new relation
FROMNODE1,FROMNODE2:RelationName1,RelationName2;TONODE1,TONODE2  = defines the relation between FROMNODES and TONODES
Both the nodes and relations must already exist.

RelationType = DIRECTED,NONDIRECTED,RECIPROCAL,COMPLEX,ATTRIBUTE

DIRECTED : a relation from the first nodes to the second
NONDIRECTED: a relation between the first to the second and the second to the first 
RECIPROCAL:  defines two relations A and B (given in Content) such that nodes1Anodes2  implies nodes2Bnodes1
COMPLEX: used to define a relation in terms of other relations.
For example: A:COMPLEX;B>C  =   A is a B of a C  (an aunt is the sister of a mother)
ATTRIBUTE: Attaches strings to a node under the name of the ATTRIBUTE ... DEFINITION is a predefined ATTRIBUTE

$ = Shows nodes
$$Node1;RelationName1,RelationName2
Finds the node by following out the relations in sequence. To find ALL the nodes related through a relation, use a * after relation name

$$$ Shows all relations 
$$$$ shows the entire knowledgebase

2) By entering the call into the text, using the preceding syntax, with the call surround in double curly brackets

3) By enter the call as a keyword, either with in the text surrounded by single curly brackets, or with the list of keywords. IN THIS CASE THE NODE is ALSO ENTERED AS A KEYWORD.
In this case, the call syntax is different, and more limited.  No lists are possible. QUESTION MARK is used before the call and between the relation and the content.  Nodes can be defined, and relations between nodes established, but relations cannot be created.

4) By entering the call in a search command.  This can only be used for "showing" with the results of the "show" being sent to the search.
e.g. ?:<?Poo?friend>  =  searches for the friend(s) of Poo rather than Poo himself.

SOME other commands attached to the generalknowledge class.

switchgeneralknowledge --- allows you to switch between the temporary knowledgebase, a universal knowledgebase, and one attached to the specific notebook the is open.

cleargeneralknowledge --- deletes everything ! Use with care!

dumpgeneralknowledge,dumpknowledge   Saves as a formatted text file that can be read back in to recreate the knowledge

loadknowledge -- Loads a dumped knowledge base.

showgeneralknowledge -- Shows the formatted text of the generalknowledge without saving it ...