Add files via upload
ADDS A NEW command 'sheet' which is similar to 'multi', creating a tiled multi-note sheet, but allows you to scroll through the a larger sheet from a small window...
Commands: arrow keys, i j k m as alternatives, and x to increase speed of scrolling, z to decrease speed of scrolling, 
and finally --- very important --- b to exist the window. 

Depends on the CURSES module, which is excruciating!