Add files via upload
Corrected the s4earch function: it is now possible to search for keyword phrases. 
Corrected an error which was making it impossible to open new files; this works for both "database" or "shelf" mode, though I also disabled the shelf mode. To un-disable, change the default of "autodatabase" to False.