complex equivalences
Extended the equivalence classes to support complex equivalences like: (evil & genius), Dr._Evil, Malvolia 

These work in both directions (from simple to complex, from complex to simple-or complex)), though the former has certain limitations. Even though I use logical deduction to establish if the complex phrase is logically implied by the search query, I haven't figured out how to "replace" the logical sense of the search phrase, rather than the term... SO just ADD the new sense with a logical "or", resulting in an inaccurate work is the search query is more complex than the substituted term.

Need to figure out how to deal with the spaces in Keywords.