fixed searching with equivalences, support ~
I simplified the complex equivalence search routine by using a simple replace to switch between AND (&) and OR (|), which is necessary for the chained entailments to function. 
To explain: equivalences are treated as a set of possible search terms, not as logical equivalence. 
Hence if (A,B,C) constitutes an equivalence class, then A will generate the search phrase A OR B OR C. 
The purpose of the search function, however, is not to generate the logical implications of this, such as ~(~A & ~B & ~C), but rather to find other complex logical phrases that are implied by the new search phrase. Hence, it is necessary to change the OR linking the terms of the search phrase, into an AND.
i.e A & B & C.

Hence, with the following equivalences having been defined:
 (A&B)=>C
(D&E)=>F
(C&F)=>G

The search phrase: A&B&D&E will first transform into C&F, and then into G. 

I also added (limited) support for ~.

Complex equivalences cannot yet contain ~. Might see if I can include this. (Got rid of it, perhaps needlessly, in the process of debugging).

Alas, anyway: asking myself, why didn't I just use OWL. This is fun in a frustrating way to program, and challenging (at least for my very limited skills), --but it is clear to me that I would have to approach it much more systematically... Even though, for my modest notetaking purposes, the limited functionality of equivalences is still very useful.