corrected major issues with the equivalence sub
An important lesson was learned: if you have to submit an object to transformations, reduce it before hand to its most logical simple, expressive form. Alas, I failed to do this at the outset, and the result was several gruelling days of debugging. Or, just think before you spend too much time solving problems that don't really exist.
I rewrote the equivalence routines for the search with a substitution class, which substitutes the terms of the search query for ten-digit number codes. I thought of this as a way to avoid errors that result from relying on text substitution to transform the query phrase. These result from having key words bracketed with <> but ordinary search terms not bracketed at all... And there was also a issue with multi word keyword phrases... 
I realize in retrospect that I could have resolved all these issues with a simple fix that was forced on me anyway, after days of fixing one problem and creating another.
Namely: all I needed to do was create a special bracket for the text words, so that all terms would be bracketed. As long as the bracketing expression aren't used within the phrase, replacement errors can't appear. The substitution class, and all the challenges of its implementation, was completely unnecessary...
Oh well_ A learning experience!