treats alternative name forms as equivalence
When a keyword is entered using a PERIOD between different parts of a name to automatically generate alternative forms, you will be asked if you want to add as an equivalence class.

e.g.  
Martin.Heidegger 
generates the names: M Heidegger, Martin Heidegger, Heidegger

This can be automatically registered as an equivalence class, with only Martin Heidegger kept as a keyword.