kindle notes compatibility
New  command "kindle"  which can be  used to  quickly create notes from kindle highlights.  Use the export function on the PC kindle app, and then   copy  the text from the HML,  and save as a textfile to the "textfile" folder.  Then type "kindle"  and, when asked, enter the  filename   (without .txt extension)